requirements.txt
--extra-index-url https://git-package-read-token:hiZnxRXN2S1b-BsEavXJ@git-ce.rwth-aachen.de/api/v4/projects/4473/packages/pypi/simple
s3-smart-open
requests
scikit-learn
flask
scipy>=1.0.0,<=1.4.1
numpy
pandas
absl-py
numpy
pymongo
gpyopt
parameter-sherpa